+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/domain/Diary.java
@@ -3,7 +3,7 @@ package com.startingblue.fourtooncookie.diary.domain;
 import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.diary.domain.converter.DiaryPaintingImageGenerationStatusListConverter;
 import com.startingblue.fourtooncookie.global.domain.BaseEntity;
-import com.startingblue.fourtooncookie.global.converter.jpa.UrlListToStringConverter;
+import com.startingblue.fourtooncookie.global.converter.UrlListToStringConverter;
 import jakarta.persistence.*;
 import jakarta.validation.*;
 import jakarta.validation.constraints.NotBlank;
